requirements.txt
aiogram
aiohttp
python-dotenv
cache
asyncio
